movilmedica/requirements.txt
appdirs==1.4.4
asn1crypto==1.5.1
attrs==21.4.0
Babel==2.6.0
beautifulsoup4==4.10.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
cryptography==36.0.2
dbus-python==1.2.16
decorator==4.3.0
defusedxml==0.7.1
docutils==0.14
ebaysdk==2.1.5
freezegun==0.3.15
gevent==20.9.0
greenlet==0.4.17
idna==2.6
isodate==0.6.1
Jinja2==2.11.2
libsass==0.17.0
lxml==4.6.1
Mako==1.0.7
MarkupSafe==1.1.0
num2words==0.5.6
ofxparse==0.19
passlib==1.7.1
Pillow==8.1.1
polib==1.1.0
psutil==5.6.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.1
pyOpenSSL==19.0.0
pyparsing==3.0.7
PyPDF2==1.26.0
pyserial==3.4
python-dateutil==2.7.3
python-ldap==3.1.0
python-stdnum==1.8
pytz==2019.1
pyusb==1.0.2
qrcode==6.1
reportlab==3.5.55
requests==2.21.0
requests-toolbelt==0.9.1
six==1.14.0
soupsieve==2.3.1
Unidecode==1.3.4
urllib3==1.24.3
vobject==0.9.6.1
Werkzeug==0.16.1
xades==0.2.2
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
xmlsig==0.1.7
zeep==3.2.0
zope.event==4.5.0
zope.interface==5.4.0
